third_party/WebKit/LayoutTests/fast/dom/Element/scrollTop-scrollLeft-body-expected.txt

Test for scrollTop/Left and scrollingElement values of scrollable document.body in quirks and strict modes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS iframeStrict1.contentDocument.scrollingElement is iframeStrict1.contentDocument.documentElement
PASS iframeStrict1.contentDocument.documentElement.scrollTop is 0
PASS iframeStrict1.contentDocument.documentElement.scrollLeft is 100
PASS iframeStrict1.contentDocument.body.scrollTop is 200
PASS iframeStrict1.contentDocument.body.scrollLeft is 0
PASS iframeStrict1.contentWindow.pageYOffset is 0
PASS iframeStrict1.contentWindow.pageXOffset is 100
PASS iframeQuirks1.contentDocument.scrollingElement is null
PASS iframeQuirks1.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks1.contentDocument.documentElement.scrollLeft is 0
PASS iframeQuirks1.contentDocument.body.scrollTop is 200
PASS iframeQuirks1.contentDocument.body.scrollLeft is 0
PASS iframeQuirks1.contentWindow.pageYOffset is 0
PASS iframeQuirks1.contentWindow.pageXOffset is 0
PASS iframeStrict2.contentDocument.scrollingElement is iframeStrict2.contentDocument.documentElement
PASS iframeStrict2.contentDocument.documentElement.scrollTop is 0
PASS iframeStrict2.contentDocument.documentElement.scrollLeft is 100
PASS iframeStrict2.contentDocument.body.scrollTop is 0
PASS iframeStrict2.contentDocument.body.scrollLeft is 0
PASS iframeStrict2.contentWindow.pageYOffset is 0
PASS iframeStrict2.contentWindow.pageXOffset is 100
PASS iframeQuirks2.contentDocument.scrollingElement is null
PASS iframeQuirks2.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks2.contentDocument.documentElement.scrollLeft is 0
PASS iframeQuirks2.contentDocument.body.scrollTop is 0
PASS iframeQuirks2.contentDocument.body.scrollLeft is 0
PASS iframeQuirks2.contentWindow.pageYOffset is 0
PASS iframeQuirks2.contentWindow.pageXOffset is 0
PASS iframeStrict3.contentDocument.scrollingElement is iframeStrict3.contentDocument.documentElement
PASS iframeStrict3.contentDocument.documentElement.scrollTop is 0
PASS iframeStrict3.contentDocument.documentElement.scrollLeft is 100
PASS iframeStrict3.contentDocument.body.scrollTop is 200
PASS iframeStrict3.contentDocument.body.scrollLeft is 0
PASS iframeStrict3.contentWindow.pageYOffset is 0
PASS iframeStrict3.contentWindow.pageXOffset is 100
PASS iframeQuirks3.contentDocument.scrollingElement is null
PASS iframeQuirks3.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks3.contentDocument.documentElement.scrollLeft is 0
PASS iframeQuirks3.contentDocument.body.scrollTop is 200
PASS iframeQuirks3.contentDocument.body.scrollLeft is 0
PASS iframeQuirks3.contentWindow.pageYOffset is 0
PASS iframeQuirks3.contentWindow.pageXOffset is 0
PASS iframeStrict4.contentDocument.scrollingElement is iframeStrict4.contentDocument.documentElement
PASS iframeStrict4.contentDocument.body.scrollTop is 0
PASS iframeStrict4.contentDocument.body.scrollLeft is 0
PASS iframeQuirks4.contentDocument.scrollingElement is iframeQuirks4.contentDocument.body
PASS iframeQuirks4.contentDocument.body.scrollTop is 4000
PASS iframeQuirks4.contentDocument.body.scrollLeft is 3000
PASS successfullyParsed is true

TEST COMPLETE

       
